feature: add passive logic and specify skill logic for unit

+++ Assets/_Project/_Scripts/Configs/PassiveConfigs/WolfLeaderRoarPassive.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "WolfLeaderRoar", menuName = "Configs/Passive/WolfLeaderRoar")]
 public class WolfLeaderRoar : PassiveConfig
 {
-    public int[] hpThreshold;
+    public float[] hpThreshold;
 }
